Update home3.html
Made sure all divs were connected with their partner; found code with an extra html tag under the About/Mission Cards comment that was making the navbar, jumbotron, and footer not reach the edges of the screen. Commented around it to quarantine it, in case we need it later for something.